[IMP] l10n_pl: merge l10n_pl_jpk into l10n_pl
Merge l10n_pl_jpk as it is a functionality that every user that uses l10n_pl should have.
Especially, since the fields will be necessary for the e-invoicing.
Also merge l10n_pl_delivery_date and delivery_date on account_move.
This means the module l10n_pl_sale_stock is not necessary anymore.
