+++ addons/l10n_pl/models/__init__.py
@@ -1,2 +1,8 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
+from . import product
+from . import account_move
+from . import res_partner
+from . import res_company
+from . import res_config_settings
+from . import l10n_pl_tax_office
 from . import template_pl
