+++ addons/l10n_pl/models/l10n_pl_tax_office.py
@@ -2,7 +2,7 @@ from odoo import fields, models
 
 
 class TaxOffice(models.Model):
-    _name = 'l10n_pl_jpk.l10n_pl_tax_office'
+    _name = 'l10n_pl.l10n_pl_tax_office'
     _description = 'Tax Office in Poland'
     _rec_names_search = ['name', 'code']
     _order = 'code'
