+++ addons/l10n_pl/models/res_config_settings.py
