+++ addons/l10n_pl/security/ir.model.access.csv
@@ -1,0 +1,2 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_l10n_pl_tax_office,l10n_pl.l10n_pl_tax_office,model_l10n_pl_l10n_pl_tax_office,account.group_account_user,1,1,1,0
