+++ addons/l10n_pl/views/account_move_views.xml
@@ -5,13 +5,6 @@
         <field name="model">account.move</field>
         <field name="inherit_id" ref="account.view_move_form"/>
         <field name="arch" type="xml">
-            <field name="invoice_date" position="after">
-                <field name="l10n_pl_show_delivery_date" invisible="1"/>
-                <field name="l10n_pl_delivery_date" string="Delivery Date" invisible="l10n_pl_show_delivery_date"/>
-            </field>
-            <field name="qr_code_method" position="after">
-                <field name="l10n_pl_delivery_date" string="Delivery Date"/>
-            </field>
             <notebook position="inside">
                 <page id="pl_extra" string="PL Extra" invisible="move_type not in ('out_invoice', 'out_refund')">
                     <group>
