+++ addons/l10n_pl/views/product_views.xml
