+++ addons/l10n_pl/views/res_config_settings_views.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
     <record id="res_config_settings_view_form" model="ir.ui.view">
-        <field name="name">res.config.settings.view.form.inherit.l10n_pl_jpk</field>
+        <field name="name">res.config.settings.view.form.inherit.l10n_pl</field>
         <field name="model">res.config.settings</field>
         <field name="inherit_id" ref="account.res_config_settings_view_form"/>
         <field name="arch" type="xml">
             <xpath expr="//block[@id='invoicing_settings']" position="after">
                 <block title="Polish Localization" invisible="country_code != 'PL'">
-                    <div id="l10n_pl_jpk_section" class="row mt16 o_settings_container">
+                    <div id="l10n_pl_section" class="row mt16 o_settings_container">
                         <div class="col-12 col-lg-6 o_setting_box">
                             <div class="o_setting_left_pane"/>
                             <div class="o_setting_right_pane">
