+++ addons/l10n_pl/views/res_partner_views.xml
@@ -2,8 +2,8 @@
 <odoo>
     <data>
 
-        <record  model="ir.ui.view" id="res_partner_account_pl_jpk_form">
-            <field name="name">res.partner.account.pl.jpk.form</field>
+        <record  model="ir.ui.view" id="res_partner_account_pl_form">
+            <field name="name">res.partner.account.pl.form</field>
             <field name="model">res.partner</field>
             <field name="inherit_id" ref="account.view_partner_property_form"/>
             <field name="arch" type="xml">
