--- addons/l10n_pl_jpk/models/res_company.py
@@ -1,7 +1,0 @@
-from odoo import fields, models
-
-
-class Company(models.Model):
-    _inherit = 'res.company'
-
-    l10n_pl_reports_tax_office_id = fields.Many2one('l10n_pl_jpk.l10n_pl_tax_office', string='Tax Office')
