--- addons/l10n_pl_sale_stock/__init__.py
@@ -1,2 +1,0 @@
-
-from . import models
