--- addons/l10n_pl_sale_stock/models/__init__.py
@@ -1,2 +1,0 @@
-
-from . import account_move
